filtro para eliminar provincias

+++ app/Http/Controllers/ProvinciaController.php
@@ -176,12 +176,18 @@ class ProvinciaController extends Controller
       $_info = $provincia->codigo.' '.$provincia->nombre;
       $deptos = count($provincia->departamentos);
       if ($deptos == 0){
-        if ($provincia->delete()) {
-              Log::info('Se borró la provincia: '.$_info);
-              $respuesta = ['statusCode'=> 200,'message' => 'Se eliminó la provincia: '.$_info];
-        }else{
-              Log::error('NO se borró la provincia: '.$_info);
-              $respuesta = ['statusCode'=> 304,'message' => 'NO se eliminó la provincia: '.$_info];
+        if (Auth::user()->can($provincia->codigo, 'web')) { //podría ser tambien provincia->nombre (definir estandar)
+            Log::info('El usuario posee el filtro: '.$provincia->codigo);
+            if ($provincia->delete()) {
+                Log::info('Se borró la provincia: '.$_info);
+                $respuesta = ['statusCode'=> 200,'message' => 'Se eliminó la provincia: '.$_info];
+            }else{
+                    Log::error('NO se borró la provincia: '.$_info);
+                    $respuesta = ['statusCode'=> 304,'message' => 'NO se eliminó la provincia: '.$_info];
+            }
+        } else {
+            Log::error('El usuario no posee el filtro: '.$provincia->codigo);
+            $respuesta = ['statusCode'=> 304,'message' => 'No cuenta con la autorización para eliminar la provincia: '.$_info];
         }
       } else {
           $respuesta = ['statusCode'=> 304,'message' => 'Existen '.$deptos.' departamentos que dependen de ésta provincia. '.$_info];
